splittet tjeneste tester i to, en for valg og en for brevet
