+++ formidling/src/test/java/no/nav/ung/sak/formidling/informasjonsbrev/InformasjonsbrevTjenesteBrevTest.java
@@ -1,32 +1,30 @@
-package no.nav.ung.sak.formidling;
+package no.nav.ung.sak.formidling.informasjonsbrev;
 
 import jakarta.inject.Inject;
 import jakarta.persistence.EntityManager;
 import no.nav.k9.felles.testutilities.cdi.CdiAwareExtension;
 import no.nav.k9.prosesstask.impl.ProsessTaskRepositoryImpl;
 import no.nav.k9.prosesstask.impl.ProsessTaskTjenesteImpl;
-import no.nav.ung.kodeverk.behandling.BehandlingResultatType;
 import no.nav.ung.kodeverk.behandling.BehandlingType;
 import no.nav.ung.kodeverk.dokument.DokumentMalType;
 import no.nav.ung.kodeverk.formidling.IdType;
 import no.nav.ung.kodeverk.formidling.InformasjonsbrevMalType;
 import no.nav.ung.kodeverk.formidling.TemplateType;
-import no.nav.ung.kodeverk.formidling.UtilgjengeligÅrsak;
 import no.nav.ung.sak.behandlingslager.behandling.Behandling;
 import no.nav.ung.sak.behandlingslager.formidling.bestilling.BrevbestillingRepository;
 import no.nav.ung.sak.behandlingslager.formidling.bestilling.BrevbestillingStatusType;
 import no.nav.ung.sak.db.util.JpaExtension;
 import no.nav.ung.sak.domene.person.pdl.AktørTjeneste;
+import no.nav.ung.sak.formidling.BrevScenarioer;
+import no.nav.ung.sak.formidling.BrevTestUtils;
+import no.nav.ung.sak.formidling.GenerertBrev;
+import no.nav.ung.sak.formidling.PdlKlientFake;
 import no.nav.ung.sak.formidling.bestilling.BrevbestillingTjeneste;
 import no.nav.ung.sak.formidling.dokarkiv.DokArkivKlientFake;
-import no.nav.ung.sak.formidling.informasjonsbrev.InformasjonsbrevGenerererTjeneste;
-import no.nav.ung.sak.formidling.informasjonsbrev.InformasjonsbrevTjeneste;
 import no.nav.ung.sak.formidling.mottaker.BrevMottakerTjeneste;
 import no.nav.ung.sak.formidling.pdfgen.PdfGenKlient;
 import no.nav.ung.sak.kontrakt.formidling.informasjonsbrev.GenereltFritekstBrevDto;
 import no.nav.ung.sak.kontrakt.formidling.informasjonsbrev.InformasjonsbrevBestillingDto;
-import no.nav.ung.sak.kontrakt.formidling.informasjonsbrev.InformasjonsbrevMottakerValgDto;
-import no.nav.ung.sak.kontrakt.formidling.informasjonsbrev.InformasjonsbrevValgDto;
 import no.nav.ung.sak.test.util.UngTestRepositories;
 import no.nav.ung.sak.test.util.behandling.TestScenarioBuilder;
 import no.nav.ung.sak.test.util.behandling.UngTestScenario;
@@ -35,14 +33,13 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import java.time.LocalDate;
-import java.util.List;
 
 import static no.nav.ung.sak.formidling.HtmlAssert.assertThatHtml;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @ExtendWith(CdiAwareExtension.class)
 @ExtendWith(JpaExtension.class)
-class InformasjonsbrevTjenesteTest {
+class InformasjonsbrevTjenesteBrevTest {
 
     @Inject
     private EntityManager entityManager;
@@ -72,72 +69,6 @@ class InformasjonsbrevTjenesteTest {
                 new BrevMottakerTjeneste(new AktørTjeneste(pdlKlient),
                     ungTestRepositories.repositoryProvider().getPersonopplysningRepository())),
             brevbestillingTjeneste
-        );
-    }
-
-    @Test
-    void skal_få_generelt_fritekstbrev_med_riktige_valg() {
-        // Given
-        LocalDate fom = LocalDate.of(2024, 12, 1);
-        Behandling behandling = lagStandardBehandling(BrevScenarioer.innvilget19år(fom));
-
-        // When
-        List<InformasjonsbrevValgDto> informasjonsbrevValg = informasjonsbrevTjeneste.informasjonsbrevValg(behandling.getId());
-
-        // Then
-        assertThat(informasjonsbrevValg.size()).isEqualTo(1);
-        InformasjonsbrevValgDto first = informasjonsbrevValg.getFirst();
-        assertThat(first.malType()).isEqualTo(InformasjonsbrevMalType.GENERELT_FRITEKSTBREV);
-
-        assertThat(first.mottakere()).isEqualTo(List.of(new InformasjonsbrevMottakerValgDto(
-            behandling.getFagsak().getAktørId().getId(),
-            IdType.AKTØRID, null))
-        );
-
-        assertThat(first.støtterFritekst()).isFalse();
-        assertThat(first.støtterTittelOgFritekst()).isTrue();
-        assertThat(first.støtterTredjepartsMottaker()).isFalse();
-    }
-
-
-    @Test
-    void skal_få_generelt_fritekstbrev_på_avsluttet_behandling() {
-        // Given
-        LocalDate fom = LocalDate.of(2024, 12, 1);
-        Behandling behandling = lagStandardBehandling(BrevScenarioer.innvilget19år(fom));
-
-        behandling.setBehandlingResultatType(BehandlingResultatType.INNVILGET);
-        behandling.avsluttBehandling();
-
-        // When
-        List<InformasjonsbrevValgDto> informasjonsbrevValg = informasjonsbrevTjeneste.informasjonsbrevValg(behandling.getId());
-
-        // Then
-        assertThat(informasjonsbrevValg.size()).isEqualTo(1);
-        InformasjonsbrevValgDto first = informasjonsbrevValg.getFirst();
-        assertThat(first.malType()).isEqualTo(InformasjonsbrevMalType.GENERELT_FRITEKSTBREV);
-    }
-
-    @Test
-    void skal_få_utilgjegelig_mottaker_hvis_død() {
-        // Given
-        LocalDate fom = LocalDate.of(2024, 12, 1);
-        UngTestScenario scenario = BrevScenarioer.død19år(fom);
-
-        Behandling behandling = TestScenarioBuilder.builderMedSøknad()
-            .medBehandlingType(BehandlingType.FØRSTEGANGSSØKNAD)
-            .medUngTestGrunnlag(scenario)
-            .buildOgLagreMedUng(ungTestRepositories);
-
-        // When
-        List<InformasjonsbrevValgDto> informasjonsbrevValg = informasjonsbrevTjeneste.informasjonsbrevValg(behandling.getId());
-
-        // Then
-        assertThat(informasjonsbrevValg.size()).isEqualTo(1);
-        InformasjonsbrevValgDto first = informasjonsbrevValg.getFirst();
-        assertThat(first.mottakere()).isEqualTo(List.of(new InformasjonsbrevMottakerValgDto(
-            behandling.getFagsak().getAktørId().getId(),
-            IdType.AKTØRID, UtilgjengeligÅrsak.PERSON_DØD))
         );
     }
