+++ formidling/src/test/java/no/nav/ung/sak/formidling/informasjonsbrev/InformasjonsbrevVerifikasjon.java
@@ -1,4 +1,7 @@
-package no.nav.ung.sak.formidling;
+package no.nav.ung.sak.formidling.informasjonsbrev;
+
+import no.nav.ung.sak.formidling.BrevScenarioer;
+import no.nav.ung.sak.formidling.BrevTestUtils;
 
 import java.time.LocalDate;
